Main: Introduce crypto utilities for handling SSM files, bugref:9955 [build fix]
git-svn-id: https://www.virtualbox.org/svn/vbox/trunk@94779 cfe28804-0f27-0410-a406-dd0f0b0b656f

+++ src/VBox/Main/src-all/CryptoUtils.cpp
@@ -56,7 +56,7 @@ DECLCALLBACK(uint64_t) SsmStream::i_ssmCryptoTell(void *pvUser)
 {
     SsmStream *pThis = static_cast<SsmStream *>(pvUser);
 
-    return RTVfsFileTell(pThis->m_hVfsFile);
+    return (uint64_t)RTVfsFileTell(pThis->m_hVfsFile);
 }
